requirements.txt
django==2.2
gunicorn
django-heroku
django-import-export==1.2.0

pytz==2019.3
sqlparse==0.3.0
Pillow==6.2.0

dj-database-url==0.4.2
psycopg2==2.7.3.2
whitenoise==3.3.1
